% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ae_sample}
\alias{ae_sample}
\title{Determines the mean absolute error of an incidence forecast from predictive
Monte-Carlo samples as difference between the mean and the data}
\usage{
ae_sample(y, dat)
}
\value{
data frame with mae by date
}
\description{
Determines the mean absolute error of an incidence forecast from predictive
Monte-Carlo samples as difference between the mean and the data
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
