% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{apply_forecast_metric}
\alias{apply_forecast_metric}
\title{Determines calibration of an incidence forecast with an Anderson-Darling test of the randomised PIT histogram.}
\usage{
apply_forecast_metric(x, func, ...)
}
\arguments{
\item{x}{a data frame with 'date', 'last_obs', (observed) 'incidence',  (predicted) 'cases'and 'sample_id' columns}

\item{func}{function for forecast assessment}
}
\value{
mean and standard deviation of the resulting p values
}
\description{
Determines calibration of an incidence forecast with an Anderson-Darling test of the randomised PIT histogram.
}
\seealso{
calibration sharpness bias
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
