% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_all_forecasts.r
\name{assess_all_forecasts}
\alias{assess_all_forecasts}
\title{Applies forecast assessments and scores to the forecasts and null models for Ebola in Western Area, Sierra Leone}
\usage{
assess_all_forecasts(max_horizon)
}
\arguments{
\item{max_horizon}{maximum forecast horizon (in weeks)}
}
\value{
list of data frames with the forecast scores, by model and horizon
}
\description{
Applies forecast assessments and scores to the forecasts and null models for Ebola in Western Area, Sierra Leone
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
