% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{assess_incidence_forecast}
\alias{assess_incidence_forecast}
\title{Assesses incidence forecasts for the 2014--15 Ebola epidemic in Western Area in Sierra Leone from Monte-Carlo samples}
\usage{
assess_incidence_forecast(x, func, ...)
}
\arguments{
\item{x}{a forecast data frame with columns 'date', 'last_obs', 'incidence', 'cases' and 'sample_id', and possibly more columns (which will be grouped by)}

\item{func}{function for forecast assessment}
}
\value{
data frame with (random) PIT p-values
}
\description{
Assesses incidence forecasts for the 2014--15 Ebola epidemic in Western Area in Sierra Leone from Monte-Carlo samples
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
