% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{bias}
\alias{bias}
\title{Determines bias of an incidence forecast from predictive Monte-Carlo samples as the proportion of predictive samples greater than the data}
\usage{
bias(y, dat)
}
\value{
data frame with bias by date
}
\description{
Determines bias of an incidence forecast from predictive Monte-Carlo samples as the proportion of predictive samples greater than the data
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
