% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate quantiles for a data frame}
\usage{
calculate_quantiles(x, quantiles)
}
\arguments{
\item{quantiles}{the quantiles to compute}
}
\value{
data frame with quantiles
}
\description{
am x a data frame with a 'value' and a 'sample_id' column
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
