% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calibration_sample}
\alias{calibration_sample}
\title{Determines sharpness of an incidence forecast with an Anderson-Darling test of the randomised PIT histogram.}
\usage{
calibration_sample(y, dat, ...)
}
\arguments{
\item{y}{vector of data (length n)}

\item{dat}{Nxn matrix of predictive samples, N being the number of Monte Carlo samples}

\item{...}{other arguments for \code{\link{pit_test_sample}}}
}
\value{
data frame with mean and standard deviation of the resulting p values
}
\description{
Determines sharpness of an incidence forecast with an Anderson-Darling test of the randomised PIT histogram.
}
\seealso{
pit_test_sample
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
