% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{ebola_wa}
\alias{ebola_wa}
\title{Ebola case data set for Western Area, Sierra Leone}
\description{
Ebola case data set for Western Area, Sierra Leone
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
