% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1.r
\name{figure1}
\alias{figure1}
\title{Plots figure 1}
\usage{
figure1(conf.levels = c(0, 0.5, 0.9))
}
\arguments{
\item{conf.levels}{Confidence levels to show}
}
\value{
plot
}
\description{
Plots figure 1
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
