% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.r
\name{figure2}
\alias{figure2}
\title{Plots figure 2}
\usage{
figure2(conf.levels = c(0, 0.5, 0.9))
}
\arguments{
\item{conf.levels}{Confidence levels to show}
}
\value{
plot
}
\description{
Plots figure 2
}
\author{
Sebastian Funk
}
