% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.r
\name{figure4}
\alias{figure4}
\title{Plots figure 4}
\usage{
figure4(max_horizon = 2)
}
\arguments{
\item{max_horizon}{maximal forecast horizon}
}
\value{
plot
}
\description{
Plots figure 4
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
