% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{incidence_forecast_calibration}
\alias{incidence_forecast_calibration}
\title{Determines calibration of incidence forecasts for the 2014--15 Ebola epidemic in Western Area in Sierra Leone}
\usage{
incidence_forecast_calibration(x)
}
\arguments{
\item{x}{a forecast data frame with columns 'date', 'last_obs', 'incidence', 'R0', 'cases' and 'sample_id'}
}
\value{
data frame with (random) PIT p-values
}
\description{
Determines calibration of incidence forecasts for the 2014--15 Ebola epidemic in Western Area in Sierra Leone
}
\author{
Sebastian Funk
}
