% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_deterministic.r
\name{null_model_deterministic}
\alias{null_model_deterministic}
\title{Creates Bayesian time-series regression null model forecasts for "Assessing the performance of real-time epidemic forecasts: A case study of the 2013-16 Ebola epidemic."  (doi: 10.1101/177451).}
\usage{
null_model_deterministic(start_forecast_date = as.Date("2014-08-24"),
  forecast_horizon = 10)
}
\arguments{
\item{start_forecast_date}{The date at which to start forecasts.}

\item{forecast_horizon}{The maximum number of weeks to forecast.}
}
\value{
A data frame containing MCMC samples from the predictive probability distribution at each time point.
}
\description{
Forecasts are generated for Western Area, Sierra Leone
}
\author{
Sebastian Funk
}
