% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pit_test_sample}
\alias{pit_test_sample}
\title{Performs an Anderson-Darling test for uniformity for a randomised PIT histogram using predictive Monte-Carlo samples}
\usage{
pit_test_sample(y, dat, J, N = 10)
}
\arguments{
\item{y}{vector of data (length n)}

\item{dat}{nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples}

\item{J}{the number of bins in the PIT histogram. If not given, will use the square root of n}

\item{N}{the number of tests to perform, each time re-randomising the PIT}
}
\value{
the n p-values from the tests
}
\description{
See Eqs. 1 and 2 in Czado, Geniting and Held (2009), Predictive Model Assessment for Count Data, Biometrics Vol. 65, No. 4 (Dec., 2009), pp. 1254-1261
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
