% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{samples_bsts}
\alias{samples_bsts}
\title{Monte Carlo samples from bsts forecasts}
\description{
Monte Carlo samples from bsts forecasts
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
