% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{samples_deterministic}
\alias{samples_deterministic}
\title{Monte Carlo samples from deterministic forecasts}
\description{
Monte Carlo samples from deterministic forecasts
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
