% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{samples_semi_mechanistic}
\alias{samples_semi_mechanistic}
\title{Monte Carlo samples from semi-mechanistic forecasts}
\description{
Monte Carlo samples from semi-mechanistic forecasts
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
