% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{samples_unfocused}
\alias{samples_unfocused}
\title{Monte Carlo samples from unfocused forecasts}
\description{
Monte Carlo samples from unfocused forecasts
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
