% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sharpness}
\alias{sharpness}
\title{Determines sharpness of an incidence forecast as the width of the prediction interval}
\usage{
sharpness(y, dat, interval)
}
\arguments{
\item{interval}{prediction interval(s) to use}
}
\value{
data frame with sharpness for each interval by date
}
\description{
Determines sharpness of an incidence forecast as the width of the prediction interval
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
