% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sharpness_sample}
\alias{sharpness_sample}
\title{Determines sharpness of an incidence forecast as the width of the prediction interval}
\usage{
sharpness_sample(y, dat)
}
\arguments{
\item{interval}{prediction interval(s) to use}
}
\value{
data frame with sharpness for each interval by date
}
\description{
Determines sharpness of an incidence forecast as the width of the prediction interval
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
