% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.r
\name{table2}
\alias{table2}
\title{Prints table 2}
\usage{
table2(max_horizon = 3)
}
\arguments{
\item{max_horizon}{maximum forecast horizon (in weeks)}
}
\value{
table in latex format
}
\description{
Prints table 2
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
