% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure3.r, R/figure4.r
\name{figure3}
\alias{figure3}
\title{Plots figure 3}
\usage{
figure3(max_horizon = 5)

figure3(max_horizon = 5)
}
\arguments{
\item{max_horizon}{maximal forecast horizon}

\item{max_horizon}{maximal forecast horizon}
}
\value{
plot

plot
}
\description{
Plots figure 3

Plots figure 3
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}

Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
